We must accept finite disappointment, but never lose infinite hope

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@
 
 int main(int argc, char *argv[])
 {
-	long result;
+	unsigned long result;
 	int i, j;
 
 	if (argc != 3)
@@ -31,7 +31,7 @@ int main(int argc, char *argv[])
 		}
 	}
 
-	result = atol(argv[1]) * atoi(argv[2]);
+	result = atol(argv[1]) * atol(argv[2]);
 	printf("%lu\n", result);
 	return (0);
 }
